refactor: new file for handle dynamic suite
